requirements.txt
Flask==2.3.2
numpy==1.24.3
Pillow==10.0.0
faiss-cpu==1.7.4
psycopg2-binary==2.9.6
python-dotenv==1.0.0
Werkzeug==2.3.6
torch==2.0.1
torchvision==0.15.2
sentence-transformers==2.2.2
opencv-python==4.8.0.74
scikit-image==0.21.0
tqdm==4.65.0
matplotlib==3.7.2
scikit-learn==1.3.0
requests==2.31.0
